# api/api.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Address defines model for Address.
type Address struct {
	City       string             `json:"city"`
	Country    string             `json:"country"`
	Id         openapi_types.UUID `json:"id"`
	PostalCode string             `json:"postalCode"`
	Region     *string            `json:"region"`
	Street     string             `json:"street"`
	Unit       *string            `json:"unit"`
}

// CreatePlaceRequest defines model for CreatePlaceRequest.
type CreatePlaceRequest = Place

// PatchResource defines model for PatchResource.
type PatchResource struct {
	Op    *string                 `json:"op,omitempty"`
	Path  *string                 `json:"path,omitempty"`
	Value *map[string]interface{} `json:"value,omitempty"`
}

// Place defines model for Place.
type Place struct {
	Address     *Address            `json:"address,omitempty"`
	Description *string             `json:"description"`
	Id          *openapi_types.UUID `json:"id,omitempty"`
	Name        *string             `json:"name,omitempty"`
	Phone       *string             `json:"phone"`
	Website     *string             `json:"website"`
}

// Id defines model for id.
type Id = string

// CreatePlaceJSONRequestBody defines body for CreatePlace for application/json ContentType.
type CreatePlaceJSONRequestBody = CreatePlaceRequest

// PatchPlaceJSONRequestBody defines body for PatchPlace for application/json ContentType.
type PatchPlaceJSONRequestBody = PatchResource

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List places
	// (GET /places)
	ListPlaces(w http.ResponseWriter, r *http.Request)
	// Create an place
	// (POST /places)
	CreatePlace(w http.ResponseWriter, r *http.Request)
	// Delete place
	// (DELETE /places/{id})
	DeletePlace(w http.ResponseWriter, r *http.Request, id Id)
	// Get a place by id
	// (GET /places/{id})
	GetPlace(w http.ResponseWriter, r *http.Request, id Id)
	// Patch place
	// (PATCH /places/{id})
	PatchPlace(w http.ResponseWriter, r *http.Request, id Id)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListPlaces operation middleware
func (siw *ServerInterfaceWrapper) ListPlaces(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListPlaces(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreatePlace operation middleware
func (siw *ServerInterfaceWrapper) CreatePlace(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreatePlace(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeletePlace operation middleware
func (siw *ServerInterfaceWrapper) DeletePlace(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeletePlace(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetPlace operation middleware
func (siw *ServerInterfaceWrapper) GetPlace(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetPlace(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PatchPlace operation middleware
func (siw *ServerInterfaceWrapper) PatchPlace(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id Id

	err = runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PatchPlace(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/places", wrapper.ListPlaces)
	m.HandleFunc("POST "+options.BaseURL+"/places", wrapper.CreatePlace)
	m.HandleFunc("DELETE "+options.BaseURL+"/places/{id}", wrapper.DeletePlace)
	m.HandleFunc("GET "+options.BaseURL+"/places/{id}", wrapper.GetPlace)
	m.HandleFunc("PATCH "+options.BaseURL+"/places/{id}", wrapper.PatchPlace)

	return m
}

type DeletedResponse struct {
}

type NotFoundResponse struct {
}

type UpdatedResponse struct {
}

type ListPlacesRequestObject struct {
}

type ListPlacesResponseObject interface {
	VisitListPlacesResponse(w http.ResponseWriter) error
}

type ListPlaces200JSONResponse struct {
	Items *[]Address `json:"items,omitempty"`
	Total *int       `json:"total,omitempty"`
}

func (response ListPlaces200JSONResponse) VisitListPlacesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type CreatePlaceRequestObject struct {
	Body *CreatePlaceJSONRequestBody
}

type CreatePlaceResponseObject interface {
	VisitCreatePlaceResponse(w http.ResponseWriter) error
}

type CreatePlace201ResponseHeaders struct {
	Location interface{}
}

type CreatePlace201JSONResponse struct {
	Body    Place
	Headers CreatePlace201ResponseHeaders
}

func (response CreatePlace201JSONResponse) VisitCreatePlaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.Header().Set("Location", fmt.Sprint(response.Headers.Location))
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response.Body)
}

type DeletePlaceRequestObject struct {
	Id Id `json:"id"`
}

type DeletePlaceResponseObject interface {
	VisitDeletePlaceResponse(w http.ResponseWriter) error
}

type DeletePlace204Response = DeletedResponse

func (response DeletePlace204Response) VisitDeletePlaceResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type DeletePlace404Response = NotFoundResponse

func (response DeletePlace404Response) VisitDeletePlaceResponse(w http.ResponseWriter) error {
	w.WriteHeader(404)
	return nil
}

type GetPlaceRequestObject struct {
	Id Id `json:"id"`
}

type GetPlaceResponseObject interface {
	VisitGetPlaceResponse(w http.ResponseWriter) error
}

type GetPlace200JSONResponse Place

func (response GetPlace200JSONResponse) VisitGetPlaceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetPlace404Response = NotFoundResponse

func (response GetPlace404Response) VisitGetPlaceResponse(w http.ResponseWriter) error {
	w.WriteHeader(404)
	return nil
}

type PatchPlaceRequestObject struct {
	Id   Id `json:"id"`
	Body *PatchPlaceJSONRequestBody
}

type PatchPlaceResponseObject interface {
	VisitPatchPlaceResponse(w http.ResponseWriter) error
}

type PatchPlace204Response = UpdatedResponse

func (response PatchPlace204Response) VisitPatchPlaceResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type PatchPlace404Response = NotFoundResponse

func (response PatchPlace404Response) VisitPatchPlaceResponse(w http.ResponseWriter) error {
	w.WriteHeader(404)
	return nil
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// List places
	// (GET /places)
	ListPlaces(ctx context.Context, request ListPlacesRequestObject) (ListPlacesResponseObject, error)
	// Create an place
	// (POST /places)
	CreatePlace(ctx context.Context, request CreatePlaceRequestObject) (CreatePlaceResponseObject, error)
	// Delete place
	// (DELETE /places/{id})
	DeletePlace(ctx context.Context, request DeletePlaceRequestObject) (DeletePlaceResponseObject, error)
	// Get a place by id
	// (GET /places/{id})
	GetPlace(ctx context.Context, request GetPlaceRequestObject) (GetPlaceResponseObject, error)
	// Patch place
	// (PATCH /places/{id})
	PatchPlace(ctx context.Context, request PatchPlaceRequestObject) (PatchPlaceResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// ListPlaces operation middleware
func (sh *strictHandler) ListPlaces(w http.ResponseWriter, r *http.Request) {
	var request ListPlacesRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ListPlaces(ctx, request.(ListPlacesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ListPlaces")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ListPlacesResponseObject); ok {
		if err := validResponse.VisitListPlacesResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// CreatePlace operation middleware
func (sh *strictHandler) CreatePlace(w http.ResponseWriter, r *http.Request) {
	var request CreatePlaceRequestObject

	var body CreatePlaceJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.CreatePlace(ctx, request.(CreatePlaceRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreatePlace")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(CreatePlaceResponseObject); ok {
		if err := validResponse.VisitCreatePlaceResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// DeletePlace operation middleware
func (sh *strictHandler) DeletePlace(w http.ResponseWriter, r *http.Request, id Id) {
	var request DeletePlaceRequestObject

	request.Id = id

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.DeletePlace(ctx, request.(DeletePlaceRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeletePlace")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(DeletePlaceResponseObject); ok {
		if err := validResponse.VisitDeletePlaceResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetPlace operation middleware
func (sh *strictHandler) GetPlace(w http.ResponseWriter, r *http.Request, id Id) {
	var request GetPlaceRequestObject

	request.Id = id

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetPlace(ctx, request.(GetPlaceRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetPlace")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetPlaceResponseObject); ok {
		if err := validResponse.VisitGetPlaceResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// PatchPlace operation middleware
func (sh *strictHandler) PatchPlace(w http.ResponseWriter, r *http.Request, id Id) {
	var request PatchPlaceRequestObject

	request.Id = id

	var body PatchPlaceJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.PatchPlace(ctx, request.(PatchPlaceRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PatchPlace")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(PatchPlaceResponseObject); ok {
		if err := validResponse.VisitPatchPlaceResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}