# api/openapi.yaml
openapi: "3.0.3"
info:
  title: KRS Backend API
  version: 1.0.0
  description: Kurse API
servers:
  - url: /
paths:
  /places:
    get:
      summary: List places
      operationId: listPlaces
      tags:
        - places
      responses:
        '200':
          description: A paged list of addresses
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 1
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Address'
    post:
      summary: Create an place
      operationId: createPlace
      tags:
        - places
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
          headers:
            Location:
              $ref: '#/components/headers/Location'
  /places/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      summary: Get a place by id
      operationId: getPlace
      tags:
        - places
      responses:
        '200':
          description: Place found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Patch place
      operationId: patchPlace
      tags:
        - places
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchResource'
      responses:
        '204':
          $ref: '#/components/responses/Updated'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete place
      operationId: deletePlace
      tags:
        - places
      responses:
        '204':
          $ref: '#/components/responses/Deleted'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    Address:
      type: object
      required:
        - id
        - street
        - city
        - postalCode
        - country
      properties:
        id:
          type: string
          format: uuid
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        street:
          type: string
          example: "123 Main St"
        unit:
          type: string
          nullable: true
          example: "Apt 4B"
        city:
          type: string
          example: "Springfield"
        region:
          type: string
          nullable: true
          example: "IL"
        postalCode:
          type: string
          example: "62704"
        country:
          type: string
          example: "US"
    Place:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "b7e23ec2-1f21-4d2a-8f7f-a6d5f1b9e8e1"
        name:
          type: string
          example: "Pasta Palace"
        description:
          type: string
          nullable: true
          example: "Cozy Italian restaurant"
        address:
          $ref: '#/components/schemas/Address'
        phone:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
    CreatePlaceRequest:
      type: object
      allOf:
        - $ref: '#/components/schemas/Place'
      required:
        - name
    PatchRequest:
      type: array
      items:
        $ref: '#/components/schemas/PatchResource'
    PatchResource:
      type: object
      properties:
        op:
          type: string
        path:
          type: string
        value:
          type: object
  responses:
    NotFound:
      description: Resource not found
    Deleted:
      description: Resource deleted
    Updated:
      description: Resource updated
  headers:
    Location:
      description: URL to the resource
      example: 'http://kurse.com/api/v1/places/acd08b48-1c57-46f7-9b78-3b6c5b6d4cca'
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string

tags:
  - name: places
    description: Operations on places